refactor: streamline Home page layout with modular components and improved animations
- Refactored the Home page to utilize modular components for grid lines, center card, and integration categories, enhancing code organization and readability.
- Updated animation handling to use a unified approach for managing phases, improving the responsiveness and visual consistency of transitions.
- Removed redundant calculations and simplified the layout logic, ensuring a more efficient rendering process.
- Enhanced comments for clarity on component interactions and animation dynamics, aiding future development.
